+++ runtime/runtime.go
@@ -26,7 +26,6 @@ import (
 	"os"
 	"path"
 	"regexp"
-	"sort"
 	"strings"
 	"sync"
 	"time"
@@ -62,7 +61,6 @@ type Runtime struct {
 
 // Mountpoints should be private to the container
 func remountPrivate(mountPoint string) error {
-
 	mounted, err := mount.Mounted(mountPoint)
 	if err != nil {
 		return err
@@ -972,29 +970,4 @@ func (runtime *Runtime) ContainerGraph() *graphdb.Database {
 
 func (runtime *Runtime) SetServer(server Server) {
 	runtime.srv = server
-}
-
-// History is a convenience type for storing a list of containers,
-// ordered by creation date.
-type History []*Container
-
-func (history *History) Len() int {
-	return len(*history)
-}
-
-func (history *History) Less(i, j int) bool {
-	containers := *history
-	return containers[j].When().Before(containers[i].When())
-}
-
-func (history *History) Swap(i, j int) {
-	containers := *history
-	tmp := containers[i]
-	containers[i] = containers[j]
-	containers[j] = tmp
-}
-
-func (history *History) Add(container *Container) {
-	*history = append(*history, container)
-	sort.Sort(history)
 }
